# openapi.json
{"openapi": "3.0.2", "info": {"title": "The ChartWizard Info API", "version": "1.0"}, "paths": {"/": {"get": {"tags": ["frontend"], "summary": "About", "operationId": "about__get", "responses": {"200": {"description": "Successful Response", "content": {"text/html": {"schema": {"type": "string"}}}}}}}, "/dashboard/": {"get": {"tags": ["frontend"], "summary": "Main", "operationId": "main_dashboard__get", "responses": {"200": {"description": "Successful Response", "content": {"text/html": {"schema": {"type": "string"}}}}}}}, "/api/users/": {"post": {"tags": ["api", "users"], "summary": "Signup", "operationId": "signup_api_users__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserCreate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserRead"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/users/me/": {"get": {"tags": ["api", "users"], "summary": "Current User", "operationId": "current_user_api_users_me__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserRead"}}}}}, "security": [{"APIKeyCookie": []}]}}, "/api/users/logout/": {"post": {"tags": ["api", "users"], "summary": "Logout User", "operationId": "logout_user_api_users_logout__post", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"SessionCookie": []}]}}, "/api/questions/": {"get": {"tags": ["api", "questions"], "summary": "Read Items", "operationId": "read_items_api_questions__get", "parameters": [{"required": false, "schema": {"title": "Skip", "type": "integer", "default": 0}, "name": "skip", "in": "query"}, {"required": false, "schema": {"title": "Limit", "type": "integer", "default": 100}, "name": "limit", "in": "query"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Read Items Api Questions  Get", "type": "array", "items": {"$ref": "#/components/schemas/Question"}}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"APIKeyCookie": []}]}, "post": {"tags": ["api", "questions"], "summary": "Ask Question", "operationId": "ask_question_api_questions__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/QuestionCreate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Question"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"APIKeyCookie": []}]}}, "/api/auth/register": {"post": {"tags": ["auth"], "summary": "Register:Register", "operationId": "register_register_api_auth_register_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserCreate"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserRead"}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorModel"}, "examples": {"REGISTER_USER_ALREADY_EXISTS": {"summary": "A user with this email already exists.", "value": {"detail": "REGISTER_USER_ALREADY_EXISTS"}}, "REGISTER_INVALID_PASSWORD": {"summary": "Password validation failed.", "value": {"detail": {"code": "REGISTER_INVALID_PASSWORD", "reason": "Password should beat least 3 characters"}}}}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/auth/forgot-password": {"post": {"tags": ["auth"], "summary": "Reset:Forgot Password", "operationId": "reset_forgot_password_auth_forgot_password_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Body_reset_forgot_password_auth_forgot_password_post"}}}, "required": true}, "responses": {"202": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/auth/reset-password": {"post": {"tags": ["auth"], "summary": "Reset:Reset Password", "operationId": "reset_reset_password_auth_reset_password_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Body_reset_reset_password_auth_reset_password_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorModel"}, "examples": {"RESET_PASSWORD_BAD_TOKEN": {"summary": "Bad or expired token.", "value": {"detail": "RESET_PASSWORD_BAD_TOKEN"}}, "RESET_PASSWORD_INVALID_PASSWORD": {"summary": "Password validation failed.", "value": {"detail": {"code": "RESET_PASSWORD_INVALID_PASSWORD", "reason": "Password should be at least 3 characters"}}}}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/auth/request-verify-token": {"post": {"tags": ["auth"], "summary": "Verify:Request-Token", "operationId": "verify_request_token_auth_request_verify_token_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Body_verify_request_token_auth_request_verify_token_post"}}}, "required": true}, "responses": {"202": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/auth/verify": {"post": {"tags": ["auth"], "summary": "Verify:Verify", "operationId": "verify_verify_auth_verify_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Body_verify_verify_auth_verify_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserRead"}}}}, "400": {"description": "Bad Request", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ErrorModel"}, "examples": {"VERIFY_USER_BAD_TOKEN": {"summary": "Bad token, not existing user ornot the e-mail currently set for the user.", "value": {"detail": "VERIFY_USER_BAD_TOKEN"}}, "VERIFY_USER_ALREADY_VERIFIED": {"summary": "The user is already verified.", "value": {"detail": "VERIFY_USER_ALREADY_VERIFIED"}}}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"Body_reset_forgot_password_auth_forgot_password_post": {"title": "Body_reset_forgot_password_auth_forgot_password_post", "required": ["email"], "type": "object", "properties": {"email": {"title": "Email", "type": "string", "format": "email"}}}, "Body_reset_reset_password_auth_reset_password_post": {"title": "Body_reset_reset_password_auth_reset_password_post", "required": ["token", "password"], "type": "object", "properties": {"token": {"title": "Token", "type": "string"}, "password": {"title": "Password", "type": "string"}}}, "Body_verify_request_token_auth_request_verify_token_post": {"title": "Body_verify_request_token_auth_request_verify_token_post", "required": ["email"], "type": "object", "properties": {"email": {"title": "Email", "type": "string", "format": "email"}}}, "Body_verify_verify_auth_verify_post": {"title": "Body_verify_verify_auth_verify_post", "required": ["token"], "type": "object", "properties": {"token": {"title": "Token", "type": "string"}}}, "ErrorModel": {"title": "ErrorModel", "required": ["detail"], "type": "object", "properties": {"detail": {"title": "Detail", "anyOf": [{"type": "string"}, {"type": "object", "additionalProperties": {"type": "string"}}]}}}, "HTTPValidationError": {"title": "HTTPValidationError", "type": "object", "properties": {"detail": {"title": "Detail", "type": "array", "items": {"$ref": "#/components/schemas/ValidationError"}}}}, "Question": {"title": "Question", "required": ["user_question", "metabase_database_id", "created_by_id", "id"], "type": "object", "properties": {"user_question": {"title": "User Question", "type": "string"}, "metabase_database_id": {"title": "Metabase Database Id", "type": "string"}, "result_query": {"title": "Result Query", "type": "string"}, "created_by_id": {"title": "Created By Id", "type": "integer"}, "id": {"title": "Id", "type": "integer"}}}, "QuestionCreate": {"title": "QuestionCreate", "required": ["user_question", "metabase_database_id", "created_by_id"], "type": "object", "properties": {"user_question": {"title": "User Question", "type": "string"}, "metabase_database_id": {"title": "Metabase Database Id", "type": "string"}, "result_query": {"title": "Result Query", "type": "string"}, "created_by_id": {"title": "Created By Id", "type": "integer"}}}, "UserCreate": {"title": "UserCreate", "required": ["email", "metabase_url", "metabase_password"], "type": "object", "properties": {"email": {"title": "Email", "type": "string"}, "metabase_url": {"title": "Metabase Url", "type": "string"}, "metabase_password": {"title": "Metabase Password", "type": "string"}, "access_token": {"title": "Access Token", "type": "string"}}}, "UserRead": {"title": "UserRead", "required": ["email", "metabase_url", "access_token", "id"], "type": "object", "properties": {"email": {"title": "Email", "type": "string"}, "metabase_url": {"title": "Metabase Url", "type": "string"}, "access_token": {"title": "Access Token", "type": "string"}, "id": {"title": "Id", "type": "integer"}, "questions": {"title": "Questions", "type": "array", "items": {"$ref": "#/components/schemas/Question"}, "default": []}}}, "ValidationError": {"title": "ValidationError", "required": ["loc", "msg", "type"], "type": "object", "properties": {"loc": {"title": "Location", "type": "array", "items": {"type": "string"}}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}}}, "securitySchemes": {"APIKeyCookie": {"type": "apiKey", "in": "cookie", "name": "ChartWizard"}, "SessionCookie": {"type": "apiKey", "in": "cookie", "name": "ChartWizard_cookie"}}}}